Make contract creation more transactional
Either it should fully work or not work at all. This will now
remove the contact at the end if an error occurs anywhere during
the process.

Example: email sending failure.

Before: the contact would be created successfully, but the user
would see an error message, and no emails would be sent.

After: the contact is either created and the email sent or neither.