Use canonical URLs for og:url tag
...as per spec:

"The URL of the object, which acts as the canonical URL. Usually the same URL as the page where property tags are placed. It shouldn't include any session variables, user identifying parameters, or counters. Always use the canonical URL for this tag, or likes and shares will be spread across all of the variations of the URL."

via https://developers.facebook.com/docs/sharing/opengraph/object-properties#standard