Tribes: simplify tribes/tags code by removing tags logic (#592)
- Removes tags logic (everything is just "tribes" now)
- Removes "is" / "likes" logic for tribes

Joining or leaving a tribe -API changed, new ones:
- DELETE `/api/users/memberships/:tribeId` - leave
- POST `/api/users/memberships/:tribeId` - join

Restructures frontend Angular code for Tribes to be more simple and modular.

Tags aren't used currently and if/when we'll implement it, it's better to have it separate from tribes logic. Tags and tribes were previously in same database collection and API logic.

We'll need this more simple approach for coming up features.