server/message: Do not send back messages that belongs to deleted users
Since we don't have a way to fetch messages from deleted users we won't
return those messages to the user who is checking her/his messages.
With this we don't have to mark every message sent by a deleted user as
read and as soon as we have a way to fetch those messages we can fix it
another way.

Issue #596