Update Mongo version in docker-compose.yml
When setting up the Docker environment, the "trustroots" container crashed with
the following error:

    trustroots    | MongoDB version: 3.2.21
    trustroots    | MongoDB version incompatibility! Compatible version(s): >=3.6 <=4.0
    trustroots    | [Server] Script crashed.

I updated the Mongo image version in docker-compose.yml to 3.6, since that's the
minimum supported version, which fixed it.

Note that if you've already built the container, you'll have to either upgrade
the data files or destroy the volume, otherwise the `db_1` container will crash
with exit code 62. See this for more details:
https://stackoverflow.com/questions/47850004/mongodb-shutting-down-with-code62