Refactor email templates to extend email-with-container (#987)
Almost all of `email-templates/*.html` had very much in common with the
`email-with-container.server.view.html` that the new reference-templates use.
The table-hell was the same, just the message differed. I made all these use
the new container-template.

In the case of `signup-reminder.server.view.html`, the match was not perfect.
The small-font paragraph which counts the number of remaining reminders was
placed in the middle of the two buttons (urlconfirm & urlconfirmplaintext). I
took the liberty to join this paragraph to the main one. The two buttons then
lie at the end, like in other templates.